my_agent/requirements.txt
langgraph
langchain_anthropic
tavily-python
langchain_community
langchain_openai
GitPython
